requirements.txt
numpy
torch==1.7.1+cpu
torchvision==0.8.2+cpu
opencv-python-headless
imutils
Pillow
facenet-pytorch
boto3